+++ .gitignore
@@ -1,1 +1,2 @@
 /SWP391_Project/nbproject/private/
+/SWP391_Project/dist/
