+++ SWP391_Project/DB.sql
@@ -3,7 +3,8 @@ SELECT * FROM swp_demo.users;
 
 DELETE FROM swp_demo.users;
 
-
+INSERT users (username, password, email, display_Name, is_admin, is_verify, is_active) VALUES ('admin', 'bachtk203@', 'bachnvse@gmail.com', 'bachhi', 1, 1, 1);
+INSERT users (username, password, email, display_Name, is_admin, is_verify, is_active) VALUES ('admin2', 'bachtk203@', 'bachnvse@gmail.com', 'bachhi', 0, 1, 1);
 create database swp_demo;
