+++ SWP391_Project/web/editprofile.jsp
@@ -14,11 +14,26 @@
         <link rel="icon" type="image/x-icon" href="assets/favicon.ico" />
         <!-- Bootstrap icons-->
         <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.5.0/font/bootstrap-icons.css" rel="stylesheet" />
-        <!-- Core theme CSS (includes Bootstrap)-->
-        <link href="css/styles.css" rel="stylesheet" />
+        <link href="https://fonts.googleapis.com/css?family=Montserrat:400,500,700" rel="stylesheet">
+
+        <!-- Bootstrap -->
+        <link type="text/css" rel="stylesheet" href="css/bootstrap.min.css"/>
+
+        <!-- Slick -->
+        <link type="text/css" rel="stylesheet" href="css/slick.css"/>
+        <link type="text/css" rel="stylesheet" href="css/slick-theme.css"/>
+
+        <!-- nouislider -->
+        <link type="text/css" rel="stylesheet" href="css/nouislider.min.css"/>
+
+        <!-- Font Awesome Icon -->
+        <link rel="stylesheet" href="css/font-awesome.min.css">
+        <!-- Custom stlylesheet -->
+        <link type="text/css" rel="stylesheet" href="css/style.css" />
         <style>
             .formedit {
-                max-width: 400px;
+
+                max-width: 900px;
                 margin: 0 auto;
                 background-color: #fff;
                 padding: 20px;
@@ -26,6 +41,7 @@
                 border-radius: 5px;
                 box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
                 margin-bottom: 11em;
+                margin-top: 5em;
             }
 
             label {
@@ -45,10 +61,15 @@
                 background-color: #808080;
                 color: #fff;
                 cursor: pointer;
+                text-align: center;
+
+                max-width: 250px;
             }
 
             input[type="submit"]:hover {
                 background-color: #808080;
+                max-width: 250px;
+                text-align: center;
             }
 
             .error-message {
@@ -63,55 +84,104 @@
                 margin-bottom: 10px;
             }
 
+            .nav-borders .nav-link.active {
+                color: #0061f2;
+                border-bottom-color: #0061f2;
+            }
+            .nav-borders .nav-link {
+                color: #69707a;
+                border-bottom-width: 0.125rem;
+                border-bottom-style: solid;
+                border-bottom-color: transparent;
+                padding-top: 0.5rem;
+                padding-bottom: 0.5rem;
+                padding-left: 0;
+                padding-right: 0;
+                margin-left: 1rem;
+                margin-right: 1rem;
+            }
+
         </style>
     </head>
     <body>
         <!-- Navigation-->
-        <%@include file="components/navBar.jsp" %>
+        <%@include file="components/navbar.jsp" %>
+
         <div class="container px-4 px-lg-5 mt-5">
-            <form action="EditProfile" method="post" class="formedit">
-                <h1>Edit Profile</h1>
-                <!-- Info Message -->
-
-                <!-- ID Field -->
-                <label>ID:</label>
-                <input type="number" name="id" value="${user.id}" readonly="">
-                <br>
-
-                <!-- Username Field -->
-                <label>Username:</label>
-                <input type="text" name="username" value="${user.username}" readonly="">
-                <c:if test="${not empty errorMsg1}">
-                    <p class="error-message">${errorMsg1}</p>
-                </c:if>
-                <br>
-
-                <!-- Password Field -->
+            <nav style="margin-top: 5em;">
+                <ul style="display: flex;">
+                    
+                </ul>
 
 
-                <!-- Display Name Field -->
-                <label>Display Name:</label>
-                <input type="text" name="display" value="${user.display_name}">
-                <c:if test="${not empty errorMsg3}">
-                    <p class="error-message">${errorMsg3}</p>
-                </c:if>
-                <br>
+            </nav>
+            <form action="EditProfile" method="post" class="formedit" >
+                <h1 style="text-align: center;">Information</h1>
 
-                <!-- Email Field -->
-                <label>Email:</label>
-                <input type="text" name="email" value="${user.email}">
-                <c:if test="${not empty errorMsg4}">
-                    <p class="error-message">${errorMsg4}</p>
-                </c:if>
-                <!-- is Admin Field -->
+                <div style="display: flex; justify-content: space-between; max-width: 1000px;">
+                    <!-- First Column -->
+                    <div style="width: 48%;">
+                        <!-- ID Field -->
+                        <label>ID:</label>
+                        <input type="number" name="id" value="${user.id}" readonly="">
+                        <br>
 
+                        <!-- Display Name Field -->
+                        <label>Display Name:</label>
+                        <input type="text" name="display" value="${user.display_name}">
+                        <c:if test="${not empty errorMsg3}">
+                            <p class="error-message">${errorMsg3}</p>
+                        </c:if>
+                        <br>
+                    </div>
 
-                <a href="ChangePassword" style="text-decoration: none;"> Change Password </a>
+                    <!-- Second Column -->
+                    <div style="width: 48%;">
+                        <!-- Username Field -->
+                        <label>Username:</label>
+                        <input type="text" name="username" value="${user.username}" readonly="">
+                        <c:if test="${not empty errorMsg1}">
+                            <p class="error-message">${errorMsg1}</p>
+                        </c:if>
+                        <br>
+
+                        <!-- Email Field -->
+                        <label>Email:</label>
+                        <input type="text" name="email" value="${user.email}">
+                        <c:if test="${not empty errorMsg4}">
+                            <p class="error-message">${errorMsg4}</p>
+                        </c:if>
+                        <br>
+                    </div>
+                </div>
                 <!-- Submit Button -->
-                <input type="submit" value="Update" style="border-radius: 20px; margin-top: 1em;">
+                <div style="text-align: center;">
+                    <input type="submit" value="Update" style="border-radius: 20px; margin-top: 1em;">
+                </div>
                 <label style="color: green">${requestScope.done}</label><br/>
             </form>
-            <!-- Footer-->
+
+
+            <!-- form 2-->
+            <form action="ChangePassword" method="post" class="formedit">
+                <h1 style="text-align: center;">Change Password</h1>
+                <label for="username">Username: </label>
+                <input type="text" name="username" value="${user.username}" readonly=""><br>
+                <label for="oldPassword">Old Password:</label>
+                <input type="password" name="oldPassword" required><br>
+
+                <label for="newPassword">New Password:</label>
+                <input type="password" name="newPassword" required><br>
+
+                <label for="confirmPassword">Confirm New Password:</label>
+                <input type="password" name="confirmPassword" required><br>
+                <label style="color: red">${requestScope.fail}</label><br/>
+                <label style="color: green">${requestScope.done1}</label><br/>
+                <div style="text-align: center;">
+                    <input type="submit" value="Update" style="border-radius: 20px; margin-top: 1em;">
+                </div>
+            </form>
+
         </div>
         <%@include file="components/footer.jsp" %>
         <!-- Bootstrap core JS-->
