+++ SWP391_Project/web/landingPage.html
@@ -25,7 +25,7 @@
                     <li><a href="#">Contact Us</a></li>
                 </ul>
                 <div class="buttons" >
-                    <a class="signup" id="signup-Button">Sign Up</a>
+                    
                 </div>
             </nav>
         </header>
@@ -36,7 +36,7 @@
                     Join us to use the most reputable and reliable intermediary buying and selling service
                 </p>
                 <div class="buttons">
-                    <a href="#" class="join">Join Now</a>
+                    <a id="signin-Button" class="join">Join Now</a>
                 </div>
             </div>
             <div class="img">
@@ -44,6 +44,12 @@
             </div>
         </section>
         <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
-        <script src="jscript/signup.js"></script>
+        <script src="js1/jquery.min.js"></script>
+        <script src="js1/bootstrap.min.js"></script>
+        <script src="js1/slick.min.js"></script>
+        <script src="js1/nouislider.min.js"></script>
+        <script src="js1/jquery.zoom.min.js"></script>
+        <script src="js1/main.js"></script>
+        <<script src="jscript/signin_captcha.js"></script>
     </body>
 </html>
