+++ SWP391_Project/web/signin.jsp
@@ -31,7 +31,7 @@
                         <div class="login-wrap p-0">
                             <div style="display: flex;text-align: center;margin-bottom: 3%;">
                                 <a style="height: 40px">
-                                    <button style="border-radius: 100%;width:40px;height: 40px" onclick="backHome()">
+                                    <button style="border-radius: 100%;width:40px;height: 40px" id="backHome">
                                         <i class="fas fa-home"></i>
                                     </button>
                                 </a>
@@ -62,7 +62,7 @@
                                 </div>
                             </form>
                             <p class="w-100 text-center">&mdash; Do not have an account ? &mdash;</p>
-                           <div class="social d-flex text-center" style="justify-content:center;">
+                            <div class="social d-flex text-center" style="justify-content:center;">
                                 <button class="w-100 px-2 py-2 mr-md-1" style="border-radius: 40px;" id="signup-Button">SIGN UP</button>
                             </div>
                         </div>
@@ -70,15 +70,14 @@
                 </div>
             </div>
         </section>
-
+        <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
+        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
         <script src="js/jquery.min.js"></script>
         <script src="js/popper.js"></script>
         <script src="js/bootstrap.min.js"></script>
         <script src="js/main.js"></script>
-        <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
-        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
         <script src="jscript/signup.js"></script>
-        <script src="jscript/signin_captcha.js.js"></script>
+        <script src="jscript/signin_captcha.js"></script>
 
     </body>
 </html>
